Refactor virtualization forms
